app/requirements.txt
Django==4.0
djangorestframework==3.13.1
python-decouple==3.5
django-extensions==3.1.5
ipython==7.30.1
djangorestframework-simplejwt==5.0.0
djangorestframework-api-key==2.1.0
psycopg2==2.9.2
celery==5.2.1
django-redis==5.2.0
redis==4.0.2
